+++ src/basic/App.tsx
@@ -26,6 +26,8 @@ import { useAddProduct } from "./hooks/product/useAddProduct";
 import { useUpdateProduct } from "./hooks/product/useUpdateProduct";
 import { useDeleteProduct } from "./hooks/product/useDeleteProduct";
 import { useCompleteOrder } from "./hooks/order/useCompleteOrder";
+import { useAddCoupon } from "./hooks/coupon/useAddCoupon";
+import { useDeleteCoupon } from "./hooks/coupon/useDeleteCoupon";
 
 const App = () => {
   const [products, setProducts] = useState<ProductWithUI[]>(() => {
@@ -230,29 +232,12 @@ const App = () => {
   const { addProduct } = useAddProduct(setProducts, addNotification);
   const { updateProduct } = useUpdateProduct(setProducts, addNotification);
   const { deleteProduct } = useDeleteProduct(setProducts, addNotification);
-
-  const addCoupon = useCallback(
-    (newCoupon: Coupon) => {
-      const existingCoupon = coupons.find((c) => c.code === newCoupon.code);
-      if (existingCoupon) {
-        addNotification("이미 존재하는 쿠폰 코드입니다.", "error");
-        return;
-      }
-      setCoupons((prev) => [...prev, newCoupon]);
-      addNotification("쿠폰이 추가되었습니다.", "success");
-    },
-    [coupons, addNotification]
-  );
-
-  const deleteCoupon = useCallback(
-    (couponCode: string) => {
-      setCoupons((prev) => prev.filter((c) => c.code !== couponCode));
-      if (selectedCoupon?.code === couponCode) {
-        setSelectedCoupon(null);
-      }
-      addNotification("쿠폰이 삭제되었습니다.", "success");
-    },
-    [selectedCoupon, addNotification]
+  const { addCoupon } = useAddCoupon(coupons, setCoupons, addNotification);
+  const { deleteCoupon } = useDeleteCoupon(
+    selectedCoupon,
+    setSelectedCoupon,
+    setCoupons,
+    addNotification
   );
 
   const handleProductSubmit = (e: React.FormEvent) => {
